+++ example/lib/pages/tile_loading_error_handle.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
-import 'package:latlong/latlong.dart';
+import 'package:latlong2/latlong.dart';
 
 import '../widgets/drawer.dart';
 
@@ -49,7 +49,7 @@ class _TileLoadingErrorHandleState extends State<TileLoadingErrorHandle> {
                       errorTileCallback: (Tile tile, error) {
                         if (_needLoadingError) {
                           WidgetsBinding.instance.addPostFrameCallback((_) {
-                            Scaffold.of(context).showSnackBar(SnackBar(
+                            ScaffoldMessenger.of(context).showSnackBar(SnackBar(
                               duration: Duration(seconds: 1),
                               content: Text(
                                 error.toString(),