+++ example/lib/pages/wms_tile_layer.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
-import 'package:latlong/latlong.dart';
+import 'package:latlong2/latlong.dart';
+
 import '../widgets/drawer.dart';
 
 class WMSLayerPage extends StatelessWidget {