+++ lib/src/gestures/gestures.dart
@@ -6,8 +6,9 @@ import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/src/gestures/interactive_flag.dart';
 import 'package:flutter_map/src/gestures/latlng_tween.dart';
 import 'package:flutter_map/src/map/map.dart';
-import 'package:latlong/latlong.dart';
-import 'package:positioned_tap_detector/positioned_tap_detector.dart';
+import 'package:latlong2/latlong.dart';
+import 'package:positioned_tap_detector_2/positioned_tap_detector_2.dart';
+import 'package:flutter/physics.dart';
 
 abstract class MapGestureMixin extends State<FlutterMap>
     with TickerProviderStateMixin {
@@ -491,7 +492,13 @@ abstract class MapGestureMixin extends State<FlutterMap>
 
     _flingController
       ..value = 0.0
-      ..fling(velocity: magnitude / 1000.0);
+      ..fling(
+          velocity: magnitude / 1000.0,
+          springDescription: SpringDescription.withDampingRatio(
+            mass: 1.0,
+            stiffness: 1000.0,
+            ratio: 5.0,
+          ));
   }
 
   void handleTap(TapPosition position) {